+++ includes/settings/miscellaneous.php
@@ -94,6 +94,14 @@ class Fiber_Admin_Miscellaneous{
 			'fiber-admin-miscellaneous', // page
 			'fiad_content_section' // section
 		);
+		
+		add_settings_field(
+			'disable_content_protection', // id
+			'Disable Content Protection', // title
+			array($this, 'fiad_disable_content_protection'), // callback
+			'fiber-admin-miscellaneous', // page
+			'fiad_content_section' // section
+		);
 	}
 	
 	public function fiad_section_info(){
@@ -144,6 +152,19 @@ class Fiber_Admin_Miscellaneous{
                 <input type="checkbox" name="fiad_miscellaneous[enable_svg]"
                        id="enable_svg"
                        value="yes" <?php checked(esc_attr(fiad_get_miscellaneous_option('enable_svg')), 'yes'); ?> />
+                <span class="slider round"></span>
+            </label>
+        </fieldset>
+		<?php
+	}
+	
+	public function fiad_disable_content_protection(){
+		?>
+        <fieldset>
+            <label for="disable_content_protection" class="fiber-admin-toggle">
+                <input type="checkbox" name="fiad_miscellaneous[disable_content_protection]"
+                       id="disable_content_protection"
+                       value="yes" <?php checked(esc_attr(fiad_get_miscellaneous_option('disable_content_protection')), 'yes'); ?> />
                 <span class="slider round"></span>
             </label>
         </fieldset>